+++ src/test/java/CodeTest.java
@@ -11,12 +11,12 @@ public class CodeTest {
 
   @Test
   public void testGetCode(){
-    byte[][] validCodes= {
+    int[][] validCodes= {
             {1,2,3,4,5},
             {0},
             {0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15}
     };
-    for (byte[] validCode : validCodes){
+    for (int[] validCode : validCodes){
       Code code = new Code(validCode);
       Assertions.assertArrayEquals(validCode,code.getCode());
     }
@@ -25,7 +25,7 @@ public class CodeTest {
 
   @Test
   public void testGetLength(){
-    byte[] codeReal={1,2,3,4,5};
+    int[] codeReal={1,2,3,4,5};
     Code code = new Code(codeReal);
     Assertions.assertEquals(code.getLenght(),5);
   }
@@ -36,11 +36,11 @@ public class CodeTest {
     Assertions.assertThrows(RuntimeException.class,() -> new Code(null));
 
     //The code has a negative number
-    byte[] negativeColorArray = {8,-1};
+    int[] negativeColorArray = {8,-1};
     Assertions.assertThrows(RuntimeException.class,() -> new Code(negativeColorArray));
 
     //The code has a number greater than 127
-    //byte[] bigColorArray = {8,127}; //no me deja el IDE poner un 128 a un byte
+    //int[] bigColorArray = {8,127}; //no me deja el IDE poner un 128 a un byte
     //Assertions.assertThrows(RuntimeException.class,() -> new Core.Code(bigColorArray));
   }
 
@@ -50,7 +50,7 @@ public class CodeTest {
 
     // TODO: Buscar la manera de detectar que no es un número
     //The code has a character
-    byte[][] charInColorArrays = {
+    int[][] charInColorArrays = {
             {'r'},
             {8,'a'},
             {'4',0},
@@ -59,7 +59,7 @@ public class CodeTest {
             {' '},
             {'-'}
     };
-    for (byte[] charInColorArray : charInColorArrays)
+    for (int[] charInColorArray : charInColorArrays)
       Assertions.assertThrows(RuntimeException.class,() -> new Code(charInColorArray));
 
   }
